move element service tests to own file
