+++ test/functional/test-custom-element.js
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+import * as lolex from 'lolex';
+import {AmpEvents} from '../../src/amp-events';
 import {BaseElement} from '../../src/base-element';
 import {ElementStub, setLoadingCheckForTests} from '../../src/element-stub';
 import {LOADING_ELEMENTS_, Layout} from '../../src/layout';
@@ -37,16 +39,6 @@ import {
   stubElements,
   upgradeOrRegisterElement,
 } from '../../src/custom-element';
-// TODO(@cramforce): Move tests into their own file.
-import {
-  getElementService,
-  getElementServiceIfAvailable,
-  getElementServiceForDoc,
-  getElementServiceIfAvailableForDoc,
-} from '../../src/element-service';
-import * as lolex from 'lolex';
-import {AmpEvents} from '../../src/amp-events';
-
 
 describes.realWin('CustomElement register', {amp: 1}, env => {
 
@@ -1852,7 +1844,7 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
     let doc;
     let win;
     let elem1;
-    let intervalCallback;
+    let setIntervalCallback;
 
     beforeEach(() => {
       elements = [];
@@ -1883,7 +1875,6 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
       };
       elements.push(elem1);
 
-      intervalCallback = undefined;
       win = {
         document: doc,
         Object: {
@@ -1891,7 +1882,7 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
         },
         HTMLElement,
         setInterval: callback => {
-          intervalCallback = callback;
+          setIntervalCallback = callback;
         },
         clearInterval: () => {
         },
@@ -1916,7 +1907,7 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
       expect(win.ampExtendedElements['amp-test2']).to.be.undefined;
       expect(doc.registerElement).to.be.calledOnce;
       expect(doc.registerElement.firstCall.args[0]).to.equal('amp-test1');
-      expect(intervalCallback).to.be.undefined;
+      expect(setIntervalCallback).to.be.undefined;
     });
 
     it('should repeat stubbing when body is not available', () => {
@@ -1929,7 +1920,7 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
       expect(win.ampExtendedElements['amp-test2']).to.be.undefined;
       expect(doc.registerElement).to.be.calledOnce;
       expect(doc.registerElement.firstCall.args[0]).to.equal('amp-test1');
-      expect(intervalCallback).to.exist;
+      expect(setIntervalCallback).to.exist;
 
       // Add more elements
       const elem2 = {
@@ -1942,7 +1933,7 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
       };
       elements.push(elem2);
       doc.body = {};
-      intervalCallback();
+      setIntervalCallback();
 
       expect(win.ampExtendedElements['amp-test1']).to.equal(ElementStub);
       expect(win.ampExtendedElements['amp-test2']).to.equal(ElementStub);
@@ -1982,260 +1973,6 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
       expect(registerElement.callCount > firstCallCount).to.be.true;
       expect(registerElement.getCall(registerElement.callCount - 1).args[0])
           .to.equal('amp-test2');
-    });
-
-    it('getElementService should wait for body when not available', () => {
-      doc.body = null;  // Body not available
-      let resolvedService;
-      const p1 = getElementServiceIfAvailable(win, 'e1', 'element-1')
-          .then(service => {
-            resolvedService = service;
-            return service;
-          });
-      return Promise.resolve().then(() => {
-        expect(intervalCallback).to.exist;
-        expect(resolvedService).to.be.undefined;
-
-        // Resolve body.
-        doc.body = {};
-        intervalCallback();
-        return p1;
-      }).then(service => {
-        expect(resolvedService).to.be.null;
-        expect(service).to.be.null;
-      });
-    });
-
-    it('getElementService should resolve with body when not available', () => {
-      doc.body = {};  // Body is available
-      const p1 = getElementServiceIfAvailable(win, 'e1', 'element-1');
-      return Promise.resolve().then(() => {
-        expect(intervalCallback).to.be.undefined;
-        return p1;
-      }).then(service => {
-        expect(service).to.be.null;
-      });
-    });
-
-    it('getElementService should wait for body when available', () => {
-      doc.body = null;  // Body not available
-      let resolvedService;
-      const p1 = getElementServiceIfAvailable(win, 'e1', 'element-1')
-          .then(service => {
-            resolvedService = service;
-            return service;
-          });
-      return Promise.resolve().then(() => {
-        expect(intervalCallback).to.exist;
-        expect(resolvedService).to.be.undefined;
-
-        // Resolve body.
-        markElementScheduledForTesting(win, 'element-1');
-        registerServiceBuilder(win, 'e1', function() {
-          return {str: 'fake1'};
-        });
-        doc.body = {};
-        intervalCallback();
-        return p1;
-      }).then(service => {
-        expect(resolvedService).to.deep.equal({str: 'fake1'});
-        expect(service).to.deep.equal({str: 'fake1'});
-      });
-    });
-
-    it('getElementService should resolve with body when available', () => {
-      doc.body = {};  // Body is available
-      markElementScheduledForTesting(win, 'element-1');
-      const p1 = getElementServiceIfAvailable(win, 'e1', 'element-1');
-      return Promise.resolve().then(() => {
-        expect(intervalCallback).to.be.undefined;
-        registerServiceBuilder(win, 'e1', function() {
-          return {str: 'fake1'};
-        });
-        return p1;
-      }).then(service => {
-        expect(service).to.deep.equal({str: 'fake1'});
-      });
-    });
-  });
-});
-
-
-describes.realWin('services', {
-  amp: {
-    ampdoc: 'single',
-  },
-}, env => {
-
-  beforeEach(() => {
-    resetServiceForTesting(env.win, 'e1');
-    resetScheduledElementForTesting(env.win, 'element-1');
-    resetScheduledElementForTesting(env.win, 'element-foo');
-  });
-
-  it('should be provided by element', () => {
-    markElementScheduledForTesting(env.win, 'element-1');
-    const p1 = getElementService(env.win, 'e1', 'element-1');
-    const p2 = getElementService(env.win, 'e1', 'element-1');
-
-    registerServiceBuilder(env.win, 'e1', function() {
-      return {str: 'from e1'};
-    });
-
-    return p1.then(s1 => {
-      expect(s1).to.deep.equal({str: 'from e1'});
-      return p2.then(s2 => {
-        expect(s1).to.equal(s2);
-      });
-    });
-  });
-
-  it('should fail if element is not in page.', () => {
-    markElementScheduledForTesting(env.win, 'element-foo');
-
-    return getElementService(env.win, 'e1', 'element-bar').then(() => {
-      return 'SUCCESS';
-    }, error => {
-      return 'ERROR ' + error;
-    }).then(result => {
-      expect(result).to.match(
-          /Service e1 was requested to be provided through element-bar/);
-    });
-  });
-
-  it('should be provided by element if available', () => {
-    markElementScheduledForTesting(env.win, 'element-1');
-    const p1 = getElementServiceIfAvailable(env.win, 'e1', 'element-1');
-    const p2 = getElementServiceIfAvailable(env.win, 'e2', 'not-available');
-    registerServiceBuilder(env.win, 'e1', function() {
-      return {str: 'from e1'};
-    });
-    return p1.then(s1 => {
-      expect(s1).to.deep.equal({str: 'from e1'});
-      return p2.then(s2 => {
-        expect(s2).to.be.null;
-      });
-    });
-  });
-
-  it('should be provided by element', () => {
-    markElementScheduledForTesting(env.win, 'element-1');
-    const p1 = getElementServiceForDoc(env.ampdoc, 'e1', 'element-1');
-    const p2 = getElementServiceForDoc(env.ampdoc, 'e1', 'element-1');
-
-    registerServiceBuilder(env.win, 'e1', function() {
-      return {str: 'from e1'};
-    });
-
-    return p1.then(s1 => {
-      expect(s1).to.deep.equal({str: 'from e1'});
-      return p2.then(s2 => {
-        expect(s1).to.equal(s2);
-      });
-    });
-  });
-
-  it('should fail if element is not in page.', () => {
-    markElementScheduledForTesting(env.win, 'element-foo');
-
-    return getElementServiceForDoc(env.ampdoc, 'e1', 'element-bar').then(() => {
-      return 'SUCCESS';
-    }, error => {
-      return 'ERROR ' + error;
-    }).then(result => {
-      expect(result).to.match(
-          /Service e1 was requested to be provided through element-bar/);
-    });
-  });
-
-  it('should be provided by element if available', () => {
-    markElementScheduledForTesting(env.win, 'element-1');
-    const p1 = getElementServiceIfAvailableForDoc(
-        env.ampdoc, 'e1', 'element-1');
-    const p2 = getElementServiceIfAvailableForDoc(
-        env.ampdoc, 'e2', 'not-available');
-    registerServiceBuilder(env.win, 'e1', function() {
-      return {str: 'from e1'};
-    });
-    return p1.then(s1 => {
-      expect(s1).to.deep.equal({str: 'from e1'});
-      return p2.then(s2 => {
-        expect(s2).to.be.null;
-      });
-    });
-  });
-
-  it('getElementServiceForDoc should wait for body when not available', () => {
-    let bodyResolver;
-    env.ampdoc.bodyPromise_ = new Promise(resolve => {
-      bodyResolver = resolve;
-    });
-    let resolvedService;
-    const p1 = getElementServiceIfAvailableForDoc(env.ampdoc, 'e1', 'element-1')
-        .then(service => {
-          resolvedService = service;
-          return service;
-        });
-    return Promise.resolve().then(() => {
-      expect(resolvedService).to.be.undefined;
-
-      // Resolve body.
-      bodyResolver();
-      return p1;
-    }).then(service => {
-      expect(resolvedService).to.be.null;
-      expect(service).to.be.null;
-    });
-  });
-
-  it('getElementServiceForDoc resolve w/ body when not available', () => {
-    const p1 = getElementServiceIfAvailableForDoc(
-        env.ampdoc, 'e1', 'element-1');
-    return Promise.resolve().then(() => {
-      return p1;
-    }).then(service => {
-      expect(service).to.be.null;
-    });
-  });
-
-  it('getElementServiceForDoc should wait for body when available', () => {
-    let bodyResolver;
-    env.ampdoc.bodyPromise_ = new Promise(resolve => {
-      bodyResolver = resolve;
-    });
-    let resolvedService;
-    const p1 = getElementServiceIfAvailableForDoc(env.ampdoc, 'e1', 'element-1')
-        .then(service => {
-          resolvedService = service;
-          return service;
-        });
-    return Promise.resolve().then(() => {
-      expect(resolvedService).to.be.undefined;
-
-      // Resolve body.
-      markElementScheduledForTesting(env.win, 'element-1');
-      registerServiceBuilder(env.win, 'e1', function() {
-        return {str: 'fake1'};
-      });
-      bodyResolver();
-      return p1;
-    }).then(service => {
-      expect(resolvedService).to.deep.equal({str: 'fake1'});
-      expect(service).to.deep.equal({str: 'fake1'});
-    });
-  });
-
-  it('getElementServiceForDoc should resolve with body when available', () => {
-    markElementScheduledForTesting(env.win, 'element-1');
-    const p1 = getElementServiceIfAvailableForDoc(
-        env.ampdoc, 'e1', 'element-1');
-    return Promise.resolve().then(() => {
-      registerServiceBuilder(env.win, 'e1', function() {
-        return {str: 'fake1'};
-      });
-      return p1;
-    }).then(service => {
-      expect(service).to.deep.equal({str: 'fake1'});
     });
   });
 });
